align with add fuzz tests
