+++ datafusion/core/tests/fuzz_cases/aggregate_fuzz.rs
@@ -18,18 +18,15 @@
 use crate::fuzz_cases::aggregation_fuzzer::{
     AggregationFuzzerBuilder, DatasetGeneratorConfig, QueryBuilder,
 };
-use std::any::Any;
-use std::fmt::{Debug, Formatter};
 use std::pin::Pin;
-use std::sync::{Arc, Mutex};
-use std::task::{Context, Poll};
+use std::sync::Arc;
 
 use arrow::array::{
     types::Int64Type, Array, ArrayRef, AsArray, Int32Array, Int64Array, RecordBatch,
     StringArray, UInt64Array,
 };
 use arrow::compute::{concat_batches, SortOptions};
-use arrow::datatypes::{DataType, UInt64Type};
+use arrow::datatypes::DataType;
 use arrow::util::pretty::pretty_format_batches;
 use arrow_schema::{Field, Schema, SchemaRef};
 use datafusion::common::Result;
@@ -43,28 +40,25 @@ use datafusion::physical_plan::aggregates::{
 use datafusion::physical_plan::{collect, displayable, ExecutionPlan};
 use datafusion::prelude::{DataFrame, SessionConfig, SessionContext};
 use datafusion_common::tree_node::{TreeNode, TreeNodeRecursion, TreeNodeVisitor};
-use datafusion_common::{DataFusionError, HashMap};
+use datafusion_common::HashMap;
 use datafusion_common_runtime::JoinSet;
 use datafusion_functions_aggregate::sum::sum_udaf;
 use datafusion_physical_expr::expressions::{col, lit, Column};
-use datafusion_physical_expr::{EquivalenceProperties, Partitioning, PhysicalSortExpr};
+use datafusion_physical_expr::PhysicalSortExpr;
 use datafusion_physical_expr_common::sort_expr::LexOrdering;
-use datafusion_physical_plan::{
-    DisplayAs, DisplayFormatType, InputOrderMode, PlanProperties,
-};
-use futures::{Stream, StreamExt};
+use datafusion_physical_plan::InputOrderMode;
+use futures::StreamExt;
 use test_utils::{add_empty_batches, StringBatchGenerator};
 
 use super::record_batch_generator::get_supported_types_columns;
-use datafusion_datasource::source::DataSource;
+use crate::fuzz_cases::stream_exec::StreamExec;
 use datafusion_execution::memory_pool::units::{KB, MB};
 use datafusion_execution::memory_pool::{
     FairSpillPool, MemoryConsumer, MemoryReservation,
 };
 use datafusion_execution::runtime_env::RuntimeEnvBuilder;
-use datafusion_execution::{RecordBatchStream, SendableRecordBatchStream, TaskContext};
+use datafusion_execution::TaskContext;
 use datafusion_functions_aggregate::array_agg::array_agg_udaf;
-use datafusion_physical_plan::execution_plan::{Boundedness, EmissionType};
 use datafusion_physical_plan::metrics::MetricValue;
 use datafusion_physical_plan::stream::RecordBatchStreamAdapter;
 use rand::rngs::StdRng;
@@ -656,8 +650,6 @@ fn assert_spill_count_metric(
             panic!("Expected no spill but found SpillCount metric with value greater than 0.");
         }
 
-        println!("SpillCount = {}", spill_count);
-
         spill_count
     } else {
         panic!("No metrics returned from the operator; cannot verify spilling.");
@@ -771,113 +763,6 @@ async fn test_single_mode_aggregate_with_spill() -> Result<()> {
     assert_spill_count_metric(true, single_aggregate);
 
     Ok(())
-}
-
-/// A Mock ExecutionPlan that can be used for writing tests of other
-/// ExecutionPlans
-pub struct StreamExec {
-    /// the results to send back
-    stream: Mutex<Option<SendableRecordBatchStream>>,
-    /// if true (the default), sends data using a separate task to ensure the
-    /// batches are not available without this stream yielding first
-    use_task: bool,
-    cache: PlanProperties,
-}
-
-impl Debug for StreamExec {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "StreamExec")
-    }
-}
-
-impl StreamExec {
-    /// Create a new `MockExec` with a single partition that returns
-    /// the specified `Results`s.
-    ///
-    /// By default, the batches are not produced immediately (the
-    /// caller has to actually yield and another task must run) to
-    /// ensure any poll loops are correct. This behavior can be
-    /// changed with `with_use_task`
-    pub fn new(stream: SendableRecordBatchStream) -> Self {
-        let cache = Self::compute_properties(stream.schema());
-        Self {
-            stream: Mutex::new(Some(stream)),
-            use_task: true,
-            cache,
-        }
-    }
-
-    /// If `use_task` is true (the default) then the batches are sent
-    /// back using a separate task to ensure the underlying stream is
-    /// not immediately ready
-    pub fn with_use_task(mut self, use_task: bool) -> Self {
-        self.use_task = use_task;
-        self
-    }
-
-    /// This function creates the cache object that stores the plan properties such as schema, equivalence properties, ordering, partitioning, etc.
-    fn compute_properties(schema: SchemaRef) -> PlanProperties {
-        PlanProperties::new(
-            EquivalenceProperties::new(schema),
-            Partitioning::UnknownPartitioning(1),
-            EmissionType::Incremental,
-            Boundedness::Bounded,
-        )
-    }
-}
-
-impl DisplayAs for StreamExec {
-    fn fmt_as(&self, t: DisplayFormatType, f: &mut Formatter) -> std::fmt::Result {
-        match t {
-            DisplayFormatType::Default | DisplayFormatType::Verbose => {
-                write!(f, "StreamExec:")
-            }
-            DisplayFormatType::TreeRender => {
-                // TODO: collect info
-                write!(f, "")
-            }
-        }
-    }
-}
-
-impl ExecutionPlan for StreamExec {
-    fn name(&self) -> &'static str {
-        Self::static_name()
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn properties(&self) -> &PlanProperties {
-        &self.cache
-    }
-
-    fn children(&self) -> Vec<&Arc<dyn ExecutionPlan>> {
-        vec![]
-    }
-
-    fn with_new_children(
-        self: Arc<Self>,
-        _: Vec<Arc<dyn ExecutionPlan>>,
-    ) -> Result<Arc<dyn ExecutionPlan>> {
-        unimplemented!()
-    }
-
-    /// Returns a stream which yields data
-    fn execute(
-        &self,
-        partition: usize,
-        _context: Arc<TaskContext>,
-    ) -> Result<SendableRecordBatchStream> {
-        assert_eq!(partition, 0);
-
-        let stream = self.stream.lock().unwrap().take();
-
-        stream.ok_or(DataFusionError::Internal(
-            "Stream already consumed".to_string(),
-        ))
-    }
 }
 
 #[tokio::test]
@@ -1155,7 +1040,7 @@ async fn run_test_high_cardinality(args: RunTestHighCardinalityArgs) -> Result<u
     let mut number_of_groups = 0;
 
     let memory_pool = task_ctx.memory_pool();
-    let memory_consumer = MemoryConsumer::new("consume_memory_in_middle");
+    let memory_consumer = MemoryConsumer::new("mock_memory_consumer");
     let mut memory_reservation = memory_consumer.register(memory_pool);
 
     let mut index = 0;
